ci: build-yocto: cleanup sstate-cache after the build
We need to clean up the ci local sstate-cache that is not used, otherwise it will grow infinitely.
This will take up significant space that will not be used by anyone since invalid sstate-cache
artefacts will no longer be needed.

Since jobs on all machines in the matrix run simultaneously, we need some kind of lock
to ensure that only one runs at a time. We use the flock tool for that propose.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>